Improve selection in "Move" and "Move To" commands
"Move" and "Move To" work again, but the GUI still does not determine
if all parts are unselected.

How to reproduce the problem with selection:
1) Create a Dodo-coupling.
2) Click on the coupling to select it.
3) Click on "Move"-Button (with four arrows) on the OSE Piping Workbench
panel.
The corresponding part name "OSE_Coupling" appears in the "Move around"
dialog.
4) Click in the main view but outside the part.
Now no part is selected, but the part name "OSE_Coupling" is
still in the "Move around" dialog.
A warning "clearSelection: Empty document" appears.
The button "Unselect All" stops working.